+++ lib/devise/webauthn.rb
@@ -4,6 +4,8 @@ require "devise"
 require "webauthn"
 
 require_relative "webauthn/version"
+require_relative "webauthn/challenge_stores/session"
+require_relative "webauthn/challenge_store_access"
 require_relative "webauthn/engine"
 require_relative "webauthn/helpers/credentials_helper"
 require_relative "webauthn/routes"
@@ -11,6 +13,8 @@ require_relative "webauthn/url_helpers"
 
 module Devise
   module Webauthn
+    mattr_accessor :challenge_store, default: ChallengeStores::Session
+
     module Test
       autoload :AuthenticatorHelpers, "devise/webauthn/test/authenticator_helpers"
     end